feat: image module (#92)
* Add imageModule.ts file

* Add image module to Worksheet class

* Add image functionality to ImageModule

* Refactor image handling in worksheetXml.ts

* Refactor makeDrawingXml function
